Print newline in our generated code, not in the emu
Unconditionally printing a newline in 'print' instructions in the
emulator means we can't display several characters on the same line
for example. It's more flexible to add a "print char '\n'" in our
generated code, and keep the "print" instruction of the emulator as it
was before, without the newline.